test file structure update + reset form bug fix
