+++ src/components/inputs/__tests__/Input.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { Input } from '../inputs';
+import { Input } from '..';
 
 test('Input - renders correctly', () => {
   const tree = renderer.create(<Input />).toJSON();
