+++ src/components/modals/__tests__/AddItemModal.test.js
@@ -2,11 +2,11 @@ import React from 'react';
 import { Provider } from 'react-redux';
 import configureStore from 'redux-mock-store';
 import renderer from 'react-test-renderer';
-import RemoveItemModal from '../modals/RemoveItemModal';
+import AddItemModal from '../AddItemModal';
 
 const mockStore = configureStore([]);
 const initialState = {};
-const closeModalFn = () => undefined;
+const testFn = () => undefined;
 
 jest.mock('enquire.js', () => {
   return {
@@ -27,7 +27,7 @@ jest.mock('react-redux', () => {
   }
 });
 
-describe('RemoveItemModal', () => {
+describe('AddItemModal', () => {
   let store;
   let component;
   beforeEach(() => {
@@ -39,12 +39,12 @@ describe('RemoveItemModal', () => {
 
     component = renderer.create(
       <Provider store={store}>
-        <RemoveItemModal handleModal={closeModalFn} />
+        <AddItemModal handleModal={testFn} scrollToBottom={testFn} />
       </Provider>
     );
   });
 
-  it('RemoveItemModal - renders correctly', () => {
+  it('AddItemModal - renders correctly', () => {
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   })
